# Optimize AddToEntityList for end of list edgecase