When loading the game, the new Id will be greater than any ids in the list so we save a considerable amount of time by skipping searching where to add it.

On a large map I was testing with this saves 7 seconds on my machine