@ZehMatt 2025-04-21
This is an interesting find and I checked in a test project the difference, there is definitely an improvement but 7 seconds is quite a lot, with 1'000'000 entries in the vector and adding 100'000 to the end I got following:
```
Current: 115 ms
Optimized: 6 ms
```
So there is a clear win here but 7 seconds saved sounds like a lot.