@ZehMatt 2025-04-21
> @ZehMatt I'm using the park from this YouTube video https://youtu.be/yFmcTfHnopo?si=7YPr1O7WiSCMIgxn how fast does it load for you?

I'm definitely aware of this park 😄, I did 2 PRs also optimizing things because of it. I will give it a spin.