@ZehMatt 2025-04-21
It does definitely improve it for that specific park, I'm not sure why the difference is so fast between my test project and this one, interesting. I did approve the PR anyway even if it were just a couple ms thats still a win to me. Nice find.