@ZehMatt 2025-04-25
I took the liberty to clean up the PR, since there are a few places where this can benefit I moved it into a new function and added namespace for good measure.